Bound and validate large packing workloads

+++ tmol/score/cartbonded/potentials/cartbonded_pose_score.hh
@@ -178,7 +178,7 @@ struct CartBondedRotamerScoreDispatch {
           TPack<Real, 2, D>,          // V_t,
           TPack<Vec<Real, 3>, 2, D>,  // dV_dx_t,
           TPack<Int, 2, D>,           // dispatch_indices_t,
-          TPack<Int, 1, D>,           // n_output_intxns_for_rot_conn_offset,
+          TPack<int64_t, 1, D>,       // n_output_intxns_for_rot_conn_offset,
           TPack<Int, 1, D>            // rotconn_for_output_intxn,
           >;
 
@@ -222,7 +222,7 @@ struct CartBondedRotamerScoreDispatch {
       TView<Vec<Int, 3>, 1, D> cart_subgraph_type_offsets,
 
       TView<Int, 2, D> dispatch_indices,
-      TView<Int, 1, D> n_output_intxns_for_rot_conn_offset,
+      TView<int64_t, 1, D> n_output_intxns_for_rot_conn_offset,
       TView<Int, 1, D> rotconn_for_output_intxn,
 
       TView<Real, 2, D> dTdV  // nterms x n-dispatch
